refactor: split AI sync-rest anchors
