+++ .sampo/changesets/1022-split-ai-sync-rest-anchors.md
@@ -1,0 +1,7 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split AI feature sync-rest anchor patching into a focused runtime module while
+preserving compatibility exports and generated workspace patch behavior.
